src/test/java/org/macroing/geo4j/shape/sphere/Sphere3DReaderUnitTests.java
/**
 * Copyright 2022 J&#246;rgen Lundgren
 * 
 * This file is part of org.macroing.geo4j.
 * 
 * org.macroing.geo4j is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Lesser General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * 
 * org.macroing.geo4j is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Lesser General Public License for more details.
 * 
 * You should have received a copy of the GNU Lesser General Public License
 * along with org.macroing.geo4j. If not, see <http://www.gnu.org/licenses/>.
 */
package org.macroing.geo4j.shape.sphere;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.junit.jupiter.api.Assertions.assertTrue;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.DataInputStream;
import java.io.DataOutput;
import java.io.DataOutputStream;
import java.io.UncheckedIOException;

import org.junit.jupiter.api.Test;

@SuppressWarnings("static-method")
public final class Sphere3DReaderUnitTests {
	public Sphere3DReaderUnitTests() {
		
	}
	
	////////////////////////////////////////////////////////////////////////////////////////////////////
	
	@Test
	public void testIsSupported() {
		final Sphere3DReader sphere3DReader = new Sphere3DReader();
		
		assertTrue(sphere3DReader.isSupported(Sphere3D.ID));
		
		assertFalse(sphere3DReader.isSupported(0));
	}
	
	@Test
	public void testReadDataInput() {
		final Sphere3DReader sphere3DReader = new Sphere3DReader();
		
		final Sphere3D sphereA = new Sphere3D();
		
		final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
		
		final DataOutput dataOutput = new DataOutputStream(byteArrayOutputStream);
		
		sphereA.write(dataOutput);
		
		final byte[] bytes = byteArrayOutputStream.toByteArray();
		
		final Sphere3D sphereB = sphere3DReader.read(new DataInputStream(new ByteArrayInputStream(bytes)));
		
		assertEquals(sphereA, sphereB);
		
		assertThrows(IllegalArgumentException.class, () -> sphere3DReader.read(new DataInputStream(new ByteArrayInputStream(new byte[] {0, 0, 0, 0}))));
		assertThrows(NullPointerException.class, () -> sphere3DReader.read(null));
		assertThrows(UncheckedIOException.class, () -> sphere3DReader.read(new DataInputStream(new ByteArrayInputStream(new byte[] {}))));
	}
	
	@Test
	public void testReadDataInputInt() {
		final Sphere3DReader sphere3DReader = new Sphere3DReader();
		
		assertThrows(IllegalArgumentException.class, () -> sphere3DReader.read(new DataInputStream(new ByteArrayInputStream(new byte[] {})), 0));
		assertThrows(NullPointerException.class, () -> sphere3DReader.read(null, Sphere3D.ID));
	}
}